+++ smacc_sm_examples/sm_dance_bot/include/sm_dance_bot/superstates/ss_f_pattern_1.h
@@ -38,7 +38,7 @@ public:
 
     static constexpr float ray_lenght_meters() { return 2; }
     static constexpr float pitch_lenght_meters() { return 0.6; }
-    static constexpr float total_iterations() { return 2; }
+    static constexpr int total_iterations() { return 2; }
     static constexpr TDirection direction() { return TDirection::RIGHT; }
 
     int iteration_count;
