+++ smacc_sm_examples/sm_dance_bot/include/sm_dance_bot/superstates/ss_s_pattern_1.h
@@ -47,7 +47,7 @@ public:
 
     static constexpr float pitch1_lenght_meters() { return 0.6; }
     static constexpr float pitch2_lenght_meters() { return 3.2; }
-    static constexpr float total_iterations() { return 3; }
+    static constexpr int total_iterations() { return 3; }
     static constexpr TDirection direction() { return TDirection::RIGHT;}
 
     int iteration_count;
